float/double coordinates -> int coordinates

+++ src/PaooGame/Camera/Camera.java
@@ -6,35 +6,35 @@ import java.awt.geom.AffineTransform;
 
 public class Camera {
     private Rectangle bounds; // limitele vizibile ale camerei
-    private double x; //coordonata x a camerei
-    private double y; //coordonata y a camerei
-    private double scale; // factorul de scalare al camerei
+    private int x; //coordonata x a camerei
+    private int y; //coordonata y a camerei
+    private int scale; // factorul de scalare al camerei
 
     // starea camerei; poziție și dimensiune
-    public Camera(double x, double y, double width, double height) {
-        this.bounds = new Rectangle((int) x, (int) y, (int) width, (int) height);
+    public Camera(int x, int y, int width, int height) {
+        this.bounds = new Rectangle(x, y, width, height);
         this.x = x;
         this.y = y;
-        this.scale = 4.0;
+        this.scale = 4;
     }
     // metoda prin care se setează poziția camerei
-    public void setPosition(double x, double y) {
+    public void setPosition(int x, int y) {
         this.x = x;
         this.y = y;
     }
     // metoda prin care se setează factorul de scalare, daca este nevoie de el
-    public void setScale(double scale) {
+    public void setScale(int scale) {
         this.scale = scale;
     }
-    public double getScale()
+    public int getScale()
     {
         return scale;
     }
-    public double getX()
+    public int getX()
     {
         return x;
     }
-    public double getY()
+    public int getY()
     {
         return y;
     }
@@ -51,8 +51,8 @@ public class Camera {
     }
     public void updateCamera(Character target)
     {
-        this.x = target.GetX() - bounds.width / 2.0 + target.GetWidth() / 2.0;
-        this.y = target.GetY() - bounds.height / 2.0 + target.GetHeight() / 2.0;
-        bounds.setLocation((int)x ,(int)y);
+        this.x = (int)(target.GetX() - bounds.width / 2.0 + target.GetWidth() / 2.0);
+        this.y = (int)(target.GetY() - bounds.height / 2.0 + target.GetHeight() / 2.0);
+        bounds.setLocation(x ,y);
     }
 }
